pyproject.toml
[project]
name = "mandelbulbvariationsgenerator"
version = "0.1.0"
description = "A python package to generate fractal images with GLSL"
dependencies = [
    "numpy>=1.24.4",
    "Pillow>=10.0.0",
    "click>=8.1.3",
    "moderngl>=5.8.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
mvg = "mandelbulbvariationsgenerator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.3.0",
    "ruff>=0.0.286",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 120

[tool.ruff]
# The line-length counting method is different between Black and Ruff.
# See https://github.com/astral-sh/ruff/pull/3714 and https://github.com/psf/black/issues/1197.
# We currently use black to check line-length (E501), and ignore E501 in Ruff.
ignore = ["E501"]
